Beta

+++ ProfileImportExport.lua
@@ -1,11 +1,19 @@
 local AddonName, Addon = ...
 
+local LibDeflate = LibStub("LibDeflate")
+local LibSerialize = LibStub("LibSerialize")
+
 local L = Addon.L
 Addon.P = {}
 
 ActionBarsEnhancedProfilesMixin = {}
 ActionBarsEnhancedImportDialogMixin = {}
 ActionBarsEnhancedExportDialogMixin = {}
+
+local DefaultProfiles = {
+    ["ElvUI Style"] = "9z13pXrvu4DDe1vefMLfeOu2gIXungRIvtAczBbPmglfCxSbs6d3z39YUJmSZnZmlBXNi9PgFR8hqTXh5rsAIV1k)fCZ(KMOMMnwt14pswJ(kEU)AM7my8bsyp375C(oFFNZ5ozVbzSfAV1wFunVwl655w3RtRB4G7G9VZmKrw8dxPmEhVDXRz3QoEhNAiYWFAagmxPJdbVONRNpIm2IT99XTcvMxhFRW2(4ChWcALA2UyRs9NN8LNCYjKrfUFL2(2AHONW6c212UUVhbrgvlKHG1GR61kSB52vD2(TkJB0212Nui(olvVHkRdXrDStvC(C8SrqKHOvSd2wbrYWAPLJLRAABy6N1SrwZQzpcoDrWnSpUEtY4jrjRgKmq(KNiS2xZcdfg8Kjq2QBTvaoCddYuj9uXZsAboHH41SRx3PvJShdxoItrKcsPiQwfEDaJqAzxLDLyfuv5GnHBAbgrMMvOckseUYyx7qhq49CAfMtjBjPveXu3SOMqNQk30aKQuHxg2HLfedyIIQBIFAWlaM6Ef3o27fuPPxNM9w21RQTBfCyiG9G78MDtlGwL((34W7NjtgqczcPF2UjPD6awL(R5M7YzYKLy(FaSUwEHBJ3tE9ZKH1FkmjfCv)51983X2v2nLxmCiNIe382mWZiRk0cOJ0IbVpRQHzdOBZiUxUjU2246YaorcyVz2d4Jm8OrmtKATqkRy6mXdMk(rE5r6dWmHM1NlPClY5vj70nr1WP2F4ZoSiCsuIit()04qFPCGeQFoltuHYNECrCT189QbkCNjugUMNhHz4TpM5RLtD81T3bl2ePotgc(f8c)y8EjlJimiea21owSmzfCDN27qlAv6p(6p5XqldFQ60CBSKVw7GMrk057MabcfWQu3YLrS2VgwL(TBEZ9z)RVrIMQnVwMUmyWkfzR(4km60OPl8xOmhtXNyJroF0RFKjPeY2ne1EGoc(LQZd1lPVNJKxPi6BRHcp0X1jCV0pdqFge1arF2CSPg5EC6ZHOppI(c0C0xCi6GZc6mDOeRoPVmI(ki6Wi6igutdAEeD0SWmaDmeDCe9vr0jq0jZrNcrpdIoTb9S5GwwArd65GnTlfea4ZX2nfykcG52RUAzTv1wL6zAwuY2FZDVldEmn44doawbKvcDi6tWH8KNcXrtmPqmR4sbzOQH6icYaybiRXka0lWEnDbB)uaFogWLh9EQ2KFA0rva)Xx4cxw0MePoSNfwwQoVoFvlRzbSMOjG)4HVfM1WGIwMbwzxMDZVAPBr8XbbW(tA(Mkm8Ur7FSk9Kh8GhXZTvP)8HpKneavCxrGxy54j)F5loK3vhjXaJuOPKogWqf65IkV1xxmfWvPN83)tgwONPjDY89erFnh(lJBoB0aQW(Y(EDm0)UcfBY3Eur6n)PqW9lQY87RY8t3yJ9JZCVJoIvFhRtvNpATL41QW2bWzs2UaObQ9EX09SQ0CrvA(X7DVhj1VFE3D3xYH)6LUeRSjJP)yGMQrh81yRfh)0hl20KAqsL13rtXumlKTN(TFhNwzTTdaTP8NF5lxvSlm4neRv9S5vp9RXDj(ulWbjdmOMaX3oLuW2q8vwXPH)2D)u9NtRG(hOb9FF(5vqxQl8VEsZvjlWxVmpm1(V)",
+    ["CooldownManager Style"] = "Lr1spnruu4PoO1YQA5v0edM4wyHIk6cz0wEmgTucfiYkNPT32zsh69M5H16gmSY1nedln(pGMWEclmU8sjMWsO(yPO1Fa69XCVmJ2KMmZz((oNVZ39CUjwdnA2GQvFszyJCqOtfyZgRzdAcCFx3cfwEwGNPFGRPpWQnfsXYMoa1U0hZB6vN)A7CZMFUgMLCag9xeUQhicl0O5cCDbn8fuwb8kYNi5OqH8rXLMWlwAToimc)TEl4alz6ue47B3OM3BNOBwZY1R4cr5Goq35ZyQRD0NxzFfLezCtKPwImLs0TOpbJxHQv9a(ppbwvxRF1QgkkkOmHYIJyjODd)uD1H(1bTeW7rQppcRcgORfJZYahtF7xc4Cz6xyGC8D44eYtnJRAMAQzkPIh0OtKeh7BOrclsolq56GkH214XAL1t0UytBeOO9RbVaNYqAYcljK1q9jIkkrJ(Zdf6NzR6A)EQPEeXZ6gt7m38uffYxuKUjAir)tRDyjskJksaTcTxYXS1Kzjb6h(1NbHiYjyZuOHIyum1qAcd0vI5T04hSCqj76tMhuXoydM9g7a3OV84jS36jBwoGdOzr3UcyrZnOvqIxwvMBqI9uqR4UHuoNt4dC5q7j8a6ANn4dFdXCWjtk6WLCHLPD4vrP)327MOHfGc8SKhQj7qK0CvQb4wWyIAAxZYH83pe21JnxU(eNp8s5kYfDBrouy0lENoUualcD3W0rqAlr1lIgzvFBhB)wXVe4J4lWxnjR4SSZTw8aQ4lAGVKvhs8CK0cCbvmWjjJ2w4lBHtzHgMt6XK17OevvrJnNNhHITPZ)xl89nOPGLB80SmskflJ4hWypc92QSMUX5(9TcdFh(2KU2j)8xk0P3A6AFD7TFp9rYMNQQa49KRD6A40PVbdGU23ccONRYbaslqaVGlSjRJh0sW)wr4)JzMXiKFVoDixbPiqD7iO(YrhDIOk7T3(SPhPCMwQBmMHIEBGU2XhF8Fi)eOURa1P7U7(NJ6Sn3K2GutAazgNsG9t7Sd7wrIeo9Wd5sq9Vd",
+}
 
 function NewProfile_OnTextChanged(self)
     if not self:HasFocus() and self:GetText() == "" then
@@ -73,7 +81,7 @@ function ActionBarsEnhancedProfilesMixin:Init()
 
             local menuGenerator = function(_, rootDescription)
                 rootDescription:CreateTitle("Select Profile")
-                for profileName, data in pairs(Addon.P.profilesList) do
+                for index, profileName in ipairs(Addon.P.profilesOrder) do
                     local categoryID = profileName
                     local categoryName = profileName
                     rootDescription:CreateRadio(categoryName, IsSelected, OnSelect, categoryID)
@@ -89,7 +97,7 @@ function ActionBarsEnhancedProfilesMixin:Init()
                 rootDescription:CreateTitle("Copy Profile")
                 local currProfile = ActionBarsEnhancedProfilesMixin:GetPlayerProfile()
 
-                for profileName, data in pairs(Addon.P.profilesList) do
+                for index, profileName in ipairs(Addon.P.profilesOrder) do
                     if profileName ~= currProfile then
                         rootDescription:CreateButton(profileName, function()
                             ActionBarsEnhancedProfilesMixin:CopyProfile(profileName, currProfile)
@@ -107,8 +115,8 @@ function ActionBarsEnhancedProfilesMixin:Init()
                 rootDescription:CreateTitle("Delete Profile")
                 local currProfile = ActionBarsEnhancedProfilesMixin:GetPlayerProfile()
 
-                for profileName, data in pairs(Addon.P.profilesList) do
-                    if profileName ~= currProfile and profileName ~= "Default" then
+                for index, profileName in ipairs(Addon.P.profilesOrder) do
+                    if profileName ~= currProfile and profileName ~= "Default" and not DefaultProfiles[profileName]  then
                         rootDescription:CreateButton(profileName, function()
                             ActionBarsEnhancedProfilesMixin:DeleteProfile(profileName)
                         end)
@@ -133,7 +141,7 @@ function Addon:GetPlayerID()
     return playerID
 end
 
-function ActionBarsEnhancedProfilesMixin:SetProfile(profileName, reload)
+function ActionBarsEnhancedProfilesMixin:SetProfile(profileName, reload, config)
     local playerID = Addon:GetPlayerID()
 
     local currentProfile = profileName
@@ -149,42 +157,89 @@ function ActionBarsEnhancedProfilesMixin:SetProfile(profileName, reload)
         profileData.FontStacksScale = 1.0
     end
 
+    Addon.C["GlobalSettings"] = {}
     for key, defaultValue in pairs(Addon.Defaults) do
-        if profileData[key] ~= nil then
-            Addon.C[key] = profileData[key]
+        if profileData["GlobalSettings"] and profileData["GlobalSettings"][key] ~= nil then
+            Addon.C["GlobalSettings"][key] = profileData["GlobalSettings"][key]
         else
-            Addon.C[key] = type(defaultValue) == "table" and CopyTable(defaultValue) or defaultValue
+            Addon.C["GlobalSettings"][key] = type(defaultValue) == "table" and CopyTable(defaultValue) or defaultValue
         end
     end
 
+    for catName, catData in pairs(profileData) do
+        if catName ~= "GlobalSettings" then
+            Addon.C[catName] = Addon.C[catName] or {}
+            
+            local targetCat = Addon.C[catName]
+
+            for key, value in pairs(catData) do
+                Addon.C[catName][key] = value
+            end
+        end
+    end
+    
     Addon.P.mapping[playerID] = currentProfile
 
     if reload then
-        StaticPopup_Show("ABE_RELOAD")
+        if not StaticPopup_Visible("ABE_RELOAD") then
+            StaticPopup_Show("ABE_RELOAD")
+        end
+    end
+end
+
+function ActionBarsEnhancedProfilesMixin:InstallDefaultPresets()
+    for profileName, profileString in pairs(DefaultProfiles) do
+        if not Addon.P.profilesList[profileName] then
+            ActionBarsEnhancedImportDialogMixin:AcceptImport(_, DefaultProfiles[profileName], profileName)
+        end
     end
 end
 
 function ActionBarsEnhancedProfilesMixin:ResetProfile()
     local profileName = self:GetPlayerProfile()
-    wipe(Addon.P.profilesList[profileName])
-    self:SetProfile(profileName, true)
+    if DefaultProfiles[profileName] then
+        self:DeleteProfile(profileName)
+        ActionBarsEnhancedImportDialogMixin:AcceptImport(_, DefaultProfiles[profileName], profileName, true)
+    else
+        wipe(Addon.P.profilesList[profileName])
+        self:SetProfile(profileName, true)
+    end
 end
 
 function ActionBarsEnhancedProfilesMixin:CreateProfile(profileName)
-    Addon.P.profilesList[profileName] = {}
-    for key, defaultValue in pairs(Addon.Defaults) do
-        if Addon.C[key] ~= defaultValue then
-            Addon.P.profilesList[profileName][key] = Addon.C[key]
-        end
+    Addon.P.profilesList[profileName] = { ["GlobalSettings"] = {} }
+
+    self:AddProfileOrder(profileName)
+end
+
+function ActionBarsEnhancedProfilesMixin:ResetCatOptions(catName)
+    local profileName = self:GetPlayerProfile()
+    if Addon.P.profilesList[profileName][catName] then
+        Addon.P.profilesList[profileName][catName] = nil
+        StaticPopup_Show("ABE_RELOAD")
     end
 end
 
 function ActionBarsEnhancedProfilesMixin:DeleteProfile(profileName)
     if Addon.P.profilesList[profileName] ~= nil then
         Addon.P.profilesList[profileName] = nil
+        self:RemoveProfileOrder(profileName)
     end
 end
 
+function ActionBarsEnhancedProfilesMixin:CopyProfileCategory(fromCatName, toCatName, reload)
+    local profileName = self:GetPlayerProfile()
+    if Addon.P.profilesList[profileName][toCatName] == nil then
+        Addon.P.profilesList[profileName][toCatName] = {}
+    end
+    wipe(Addon.P.profilesList[profileName][toCatName])
+    Addon.P.profilesList[profileName][toCatName] = CopyTable(Addon.P.profilesList[profileName][fromCatName])
+    if reload then
+        if not StaticPopup_Visible("ABE_RELOAD") then
+            StaticPopup_Show("ABE_RELOAD")
+        end
+    end
+end
 function ActionBarsEnhancedProfilesMixin:CopyProfile(fromProfileName, toProfileName)
     if Addon.P.profilesList[fromProfileName] ~= nil then
         wipe(Addon.P.profilesList[toProfileName])
@@ -213,10 +268,72 @@ function ActionBarsEnhancedProfilesMixin:NeedMigrateProfile()
     return false
 end
 
+function ActionBarsEnhancedProfilesMixin:CheckProfiles15()
+    local anyMigrated = false
+    for profileName, profileData in pairs(Addon.P.profilesList) do
+        local needMigrate = self:NeedMigrateProfile15(profileName)
+        if needMigrate then
+            anyMigrated = true
+            Addon.Print("Profile "..profileName.." need migrate to v2.0")
+            self:MigrateProfile15(profileName)
+        else
+            --Addon.Print(profileName.." profile ready for v2.0")
+        end
+        self:CheckProfilesOrer(profileName)
+    end
+    if anyMigrated then
+        table.sort(Addon.P.profilesOrder)
+    end
+end
+
+function ActionBarsEnhancedProfilesMixin:NeedMigrateProfile15(profileName)
+    return not Addon.P.profilesList[profileName]["GlobalSettings"]
+end
+
+function ActionBarsEnhancedProfilesMixin:MigrateProfile15(profileName)
+    Addon.Print("Start migrating profie "..profileName.." to v2.0")
+    if not Addon.P.profilesList[profileName]["GlobalSettings"] then
+        Addon.P.profilesList[profileName]["GlobalSettings"] = {}
+    end
+
+    for key, value in pairs(Addon.P.profilesList[profileName]) do
+        if key ~= "GlobalSettings" then
+            Addon.P.profilesList[profileName]["GlobalSettings"][key] = value
+            Addon.P.profilesList[profileName][key] = nil
+        end
+    end
+    Addon.Print(profileName.." migrated to v2.0")
+end
+
+function ActionBarsEnhancedProfilesMixin:CheckProfilesOrer(profileName)
+    if Addon.P.profilesOrder == nil then
+        Addon.P.profilesOrder = {}
+    end
+
+    if Addon.P.profilesList[profileName] then
+        tInsertUnique(Addon.P.profilesOrder, profileName)
+    else
+        tDeleteItem(Addon.P.profilesOrder, profileName)
+    end
+end
+
+function ActionBarsEnhancedProfilesMixin:AddProfileOrder(profileName)
+    table.insert(Addon.P.profilesOrder, profileName)
+end
+function ActionBarsEnhancedProfilesMixin:RemoveProfileOrder(profileName)
+    tDeleteItem(Addon.P.profilesOrder, profileName)
+end
+function ActionBarsEnhancedProfilesMixin:GetProfileOrder(profileName)
+    return tIndexOf(Addon.P.profilesOrder, profileName)
+end
+
 function ActionBarsEnhancedProfilesMixin:GetPlayerProfile()
     local playerID = Addon:GetPlayerID()
     if Addon.P.mapping == nil then
         Addon.P.mapping = {}
+    end
+    if Addon.P.profilesOrder == nil then
+        Addon.P.profilesOrder = {}
     end
     if Addon.P.mapping[playerID] == nil then
         Addon.P.mapping[playerID] = "Default"
@@ -225,7 +342,8 @@ function ActionBarsEnhancedProfilesMixin:GetPlayerProfile()
         Addon.P.profilesList = {}
     end
     if Addon.P.profilesList["Default"] == nil then
-        Addon.P.profilesList["Default"] = {}
+        Addon.P.profilesList["Default"] = { ["GlobalSettings"] = {} }
+        self:AddProfileOrder("Default")
     end
     if Addon.P.mapping[playerID] ~= "Default" then
         if Addon.P.profilesList[Addon.P.mapping[playerID]] == nil then
@@ -236,39 +354,51 @@ function ActionBarsEnhancedProfilesMixin:GetPlayerProfile()
     return Addon.P.mapping[playerID]
 end
 
-function Addon.CompressData(data, options)
-    if C_EncodingUtil then
-        local dataSerialized = C_EncodingUtil.SerializeCBOR(data)
-        if dataSerialized then
-            local dataCompressed = C_EncodingUtil.CompressString(dataSerialized, Enum.CompressionMethod.Deflate, Enum.CompressionLevel.OptimizeForSize)
-            if dataCompressed then
-                local dataEncoded = C_EncodingUtil.EncodeBase64(dataCompressed)
-                return dataEncoded
-            end
-        end
-    end
+function Addon.CompressData(data)
+    local serialized = LibSerialize:Serialize(data)
+    local compressed = LibDeflate:CompressDeflate(serialized)
+    local encoded = LibDeflate:EncodeForPrint(compressed)
+    return encoded
 end
 
-function Addon.DecompressData(data, options)
-    if C_EncodingUtil then
-        local dataCompressed = C_EncodingUtil.DecodeBase64(data)
-        if dataCompressed then
-            local dataSerialized = C_EncodingUtil.DecompressString(dataCompressed)
-            if dataSerialized then
-                local dataDecompressed = C_EncodingUtil.DeserializeCBOR(dataSerialized)
-                return dataDecompressed
-            end
-        end
+function Addon.DecompressData(data)
+    local decoded = LibDeflate:DecodeForPrint(data)
+    if not decoded then
+        Addon.Print("Cant decode string")
     end
-    return Addon.Print("Cant decompress string")
+    local decompressed = LibDeflate:DecompressDeflate(decoded)
+    if not decompressed then
+        Addon.Print("Cant decompress string")
+    end
+    local success, deserialized = LibSerialize:Deserialize(decompressed)
+
+    if success then
+        return deserialized
+    end
+    return Addon.Print("Cant deserialize string")
+end
+
+function ActionBarsEnhancedProfilesMixin:SelfTest()
+    local profileName = self:GetPlayerProfile()
+    local profile = Addon.P.profilesList[profileName]
+    local exportTbl = CopyTable(profile)
+    local encoded = Addon.CompressData(exportTbl)
+    local decoded = Addon.DecompressData(encoded)
+    
+    if decoded then
+        Addon.Print("SELF TEST GOOD")
+    else
+        Addon.Print("SELF TEST BAD")
+    end
 end
 
 function ActionBarsEnhancedProfilesMixin:ExportProfile()
     local profileName = self:GetPlayerProfile()
     if Addon.P.profilesList[profileName] then
-        local exportString = Addon.CompressData(Addon.P.profilesList[profileName])
+        local exportTbl = CopyTable(Addon.P.profilesList[profileName])
 
-        
+        local exportString = Addon.CompressData(exportTbl)
+
         if not ActionBarEnhancedExportProfile then
             local ExportProfile = CreateFrame("Frame", "ActionBarEnhancedExportProfile", ActionBarEnhancedProfilesFrame, "ActionBarsEnhancedExportDialog")
             ExportProfile:SetParent(ActionBarEnhancedProfilesFrame)
@@ -283,12 +413,23 @@ function ActionBarsEnhancedProfilesMixin:ExportProfile()
             ActionBarEnhancedExportProfile.ExportControl.ExportContainer.EditBox:SetAutoFocus(true)
             ActionBarEnhancedExportProfile:Show()
         end
-
-
     end
 end
 
-function ActionBarsEnhancedImportDialogMixin:AcceptImport(profileString, profileName)
+function ActionBarsEnhancedImportDialogMixin:HasDefaultProfiles()
+    if not Addon.P.profilesList then return false end
+    
+    local i = 0
+    for profileName, data in pairs(DefaultProfiles) do
+        i = i + 1
+        if Addon.P.profilesList[profileName] then
+            i = i - 1
+        end
+    end
+    return i == 0
+end
+
+function ActionBarsEnhancedImportDialogMixin:AcceptImport(_, profileString, profileName, shouldSet)
     if not profileString then
         profileString = ActionBarEnhancedImportProfile.ImportControl.InputContainer.EditBox:GetText()
     end
@@ -300,13 +441,19 @@ function ActionBarsEnhancedImportDialogMixin:AcceptImport(profileString, profile
             Addon.Print("Profile with this name already exists")
             return false
         else
+            
             local profileTable = Addon.DecompressData(profileString)
             if profileTable then
                 Addon.P.profilesList[profileName] = CopyTable(profileTable)
                 if ActionBarEnhancedImportProfile and ActionBarEnhancedImportProfile:IsVisible() then
                     ActionBarEnhancedImportProfile:Hide()
                 end
-                ActionBarsEnhancedProfilesMixin:SetProfile(profileName, true)
+
+                ActionBarsEnhancedProfilesMixin:AddProfileOrder(profileName)
+
+                if shouldSet then
+                    ActionBarsEnhancedProfilesMixin:SetProfile(profileName, true)
+                end
                 return true
             end
         end
@@ -325,8 +472,11 @@ end
 
 function ActionBarsEnhancedProfilesMixin:GetProfiles()
     local profileTbl = {}
-    for profileName, data in pairs(Addon.P.profilesList) do
-        table.insert(profileTbl, profileName)
+    for index, profileName in ipairs(Addon.P.profilesOrder) do
+        if Addon.P.profilesList[profileName] then
+            table.insert(profileTbl, profileName)
+        end
     end
+
     return profileTbl
 end
